Rewrite benchmarks
